move shouldSkip and BodyComposer outside of apollo-api
