--- apollo-api/src/commonMain/kotlin/com/apollographql/apollo3/api/Object.kt
@@ -1,2 +1,0 @@
-package com.apollographql.apollo3.api
-
