+++ apollo-api/src/commonMain/kotlin/com/apollographql/apollo3/api/Query.kt
@@ -1,7 +1,5 @@
 package com.apollographql.apollo3.api
 
-import okio.ByteString
-
 /**
  * Represents a GraphQL query that will be sent to the server.
  */
