+++ apollo-normalized-cache/src/commonMain/kotlin/com/apollographql/apollo3/cache/normalized/internal/ShouldSkip.kt
@@ -1,4 +1,4 @@
-package com.apollographql.apollo3.api.internal
+package com.apollographql.apollo3.cache.normalized.internal
 
 import com.apollographql.apollo3.api.ResponseField
