+++ apollo-runtime/src/main/java/com/apollographql/apollo3/internal/interceptor/ApolloServerInterceptor.kt
@@ -6,7 +6,7 @@ import com.apollographql.apollo3.api.Query
 import com.apollographql.apollo3.api.cache.http.HttpCache
 import com.apollographql.apollo3.api.cache.http.HttpCachePolicy
 import com.apollographql.apollo3.api.internal.ApolloLogger
-import com.apollographql.apollo3.api.internal.OperationRequestBodyComposer
+import com.apollographql.apollo3.OperationRequestBodyComposer
 import com.apollographql.apollo3.api.internal.json.BufferedSinkJsonWriter
 import com.apollographql.apollo3.cache.ApolloCacheHeaders
 import com.apollographql.apollo3.cache.CacheHeaders
